Simplified the prior
With preconditioning there is not really a need to "recompute" Ah and Qh by
calling a function. In fact, Ah does not change at all and Qh is only Q(1)
scaled by h. Therefore, `ibm` now directly returns A and Q, and all that is left
to do at other places in the code (where previously `A!(Ah, dt)` and `Q!(Qh,
dt)` was called, is to do `Qh = Q*dt`.

This reduces code complexity, and should also increase the speed!

It might however be necessary to think about allocations at some point.