Moved the remaining dt into the preconditioner
This implementation is strongly inspired by and closely follows the
implementation in ProbNum:
https://github.com/probabilistic-numerics/probnum/blob/6b0931f17a4e6fa37e9a1dae937ad2989797dc78/src/probnum/filtsmooth/statespace/preconditioner.py#L32

By the way, this change brings more than just convenience: Previously we
computed Q*dt, which requires the allocation of a new Q-like matrix. This is now
not necessary anymore! And its also simply faster of course to _not_ do some
operation.