First POC solving a DAEProblem
The solution handling is still a bit weird though: I'm building an ODESolution
to minimize changes, and because it still kind of makes sense here. But I will
have to look into this at some later point.
What are the reasons to return a DAESolution, vs returning an ODESolution?
Do they have to fulfill different things?