Moved the state initialization out of the cache creation
It's now implemented by updating/conditioning some zero-mean standard-deviation
initial state to measurements. Things stay the same, but the code reflects
better how this should be thought of. It also makes it easier to decide /not/ to
initialize the state, or to initialize it only up to a certain order, but that's
not yet implemented.