Change the diffusion API to `estimate_{local,global}_diffusion`
Now it is made very clear that in all cases we estimate a local diffusion that
is used for local error estimation. The global MLE is only computed if
diffusionmodel not a dynamic model!

integ.sol.diffusions now contains the local diffusions; the global one is
accumulated in integ.cache.global_diffusion, when the ODE step is successful.