New diffusionmodel features and refactoring (#95)
* Specify the diffusionmodel as a struct now, instead of a symbol

* Add settings to FixedDiffusion; Remove the MAP diffusion

* Rename some diffusion models and export them

* Use the initial diffusion for extrapolating; Add this property to the MV variant

* Rescale correctly when we have a fixed diffusion model

And add an option for not calibrating

* Make things work for the FixedMVDiffusion

It's not great, but this way we can specify a scalar that is used as the
multiple of an identity matrix for the solve; it's then calibrated later
anyways.

* Add a test for an uncalibrated time-fixed diffusion

* Initialize the diffusions with NaN

This makes it more clear that the dynamic diffusion model does not have a
"global" diffusion estimate at all. Also, it should produce errors if something
goes wrong.

* Change the diffusion API to `estimate_{local,global}_diffusion`

Now it is made very clear that in all cases we estimate a local diffusion that
is used for local error estimation. The global MLE is only computed if
diffusionmodel not a dynamic model!

integ.sol.diffusions now contains the local diffusions; the global one is
accumulated in integ.cache.global_diffusion, when the ODE step is successful.

* Make the initial diff more explicit; return NaN global dynamic diffs

* Fix failing tests

* Scale the initial covariance with the initial diffusion

* Make the diagonal diffusion codes much shorter

* Fix test: Use the `default_diffusion` for initialization

* Improve the diffusion docstrings

* Improve test/correctness.jl and apply JuliaFormatter.jl