More robust tmp_cache dispatch (#189)
EKCache is the only AbstractODEFilterCache, and so it's safe (at least for now) to extend this function to the other algorithms. The other algorithms were fine because they were using a fallback that has been eliminated by https://github.com/SciML/OrdinaryDiffEq.jl/pull/1753. It would be good to get this patch in so that the ambiguity fix PR can be fully checked.