Towards a more modular state init by adding more options
This commit adss back a ForwardDiff.jl-based state initialization that was used
in a much earlier version of this package. And in almost all cases, this is not
very performant! But the initialization should ideally be more modular, the
choice being "exact" vs "inexact", and the actual autodiff implementation in the
exact case should also be easy to swap, e.g. when we'll use TaylorDiff.jl.
Therefore we try to refactor this a bit.