Refactor the plotting in favor of a Gaussian plot recipe
Previously we had a recipe for the specific solution type that we implemented.
But for some reason, this broke du to sanges of SciML. So to not have such
issues again, let's just use their plot recipe! To do this we just needed to
implement some recipes for the `Gaussian` type that we use in the solutions, and
make sure that the solution interpolation etc is doing what it should do.